reorganised util package to maek semse
