+++ src/main/java/com/casm/acled/crawler/utils/DateUtil.java
@@ -1,8 +1,29 @@
 package com.casm.acled.crawler.utils;
 
+import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableMap;
+import com.joestelmach.natty.DateGroup;
+import com.joestelmach.natty.Parser;
+import com.mdimension.jchronic.Chronic;
+import com.mdimension.jchronic.Options;
+import com.mdimension.jchronic.tags.Pointer;
+import com.mdimension.jchronic.utils.Span;
+import org.ocpsoft.prettytime.nlp.PrettyTimeParser;
+
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
+import java.time.Instant;
+import java.time.LocalDate;
+import java.time.ZoneId;
 import java.util.Date;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.function.Function;
+import java.util.stream.Collectors;
+
+import static java.util.stream.Collectors.counting;
+import static java.util.stream.Collectors.groupingBy;
 
 public class DateUtil {
     private static final String[] timeFormats = {"HH:mm:ss", "HH:mm"};
@@ -13,6 +34,95 @@ public class DateUtil {
 
     private static final String ymd_template = "\\d{4}{sep}\\d{2}{sep}\\d{2}.*";
     private static final String dmy_template = "\\d{2}{sep}\\d{2}{sep}\\d{4}.*";
+    private static Options opts = new Options(Pointer.PointerType.PAST);
+    private static PrettyTimeParser prettyParser = new PrettyTimeParser();
+    private static Parser nattyParser = new Parser();
+    private static Map<String, String> months = ImmutableMap.<String,String>builder()
+            .put("كَانُون ٱلثَّانِي",  "January")
+            .put("شُبَاط", "February")
+            .put("آذَار" ,"March")
+            .put("نَيْسَان" , "April")
+            .put("أَيَّار"	, "May")
+            .put("حَزِيرَان" , "June")
+            .put("تَمُّوز"	, "July")
+            .put("آب" , "August")
+            .put("أَيْلُول" , "September")
+            .put("تِشْرِين ٱلْأَوَّل" , "October")
+            .put("تِشْرِين ٱلثَّانِي" , "November")
+            .put("كَانُون ٱلْأَوَّل" , "December")
+
+            .put("يناير" , "January")
+            .put("فبراير" , "February")
+            .put("مارس" , "March")
+            .put("أبريل" , "April")
+            .put("إبريل" , "April")
+            .put("مايو" , "May")
+            .put("يونيو" , "June")
+            .put("يونية", "June")
+            .put("يوليو" , "July")
+            .put("يولية" , "July")
+            .put("أغسطس", "August")
+            .put("سبتمبر", "September")
+            .put("أكتوبر" , "October")
+            .put("نوفمبر" , "November")
+            .put("ديسمبر" , "December")
+            .build();
+    private static Map<String, String> russianMonths = ImmutableMap.<String,String>builder()
+            .put("января",  "January")
+            .put("февраля", "February")
+            .put("март" ,"March")
+            .put("апреля" , "April")
+            .put("май"	, "May")
+            .put("июня" , "June")
+            .put("июля"	, "July")
+            .put("августа" , "August")
+            .put("сентября" , "September")
+            .put("октября" , "October")
+            .put("ноября" , "November")
+            .put("декабря" , "December")
+            .build();
+    private static Map<String, String> frenchMonths = ImmutableMap.<String,String>builder()
+            .put("janvier",  "January")
+            .put("février", "February")
+            .put("mars" ,"March")
+            .put("avril" , "April")
+            .put("mai"	, "May")
+            .put("juin" , "June")
+            .put("juillet"	, "July")
+            .put("août" , "August")
+            .put("septembre" , "September")
+            .put("octobre" , "October")
+            .put("novembre" , "November")
+            .put("décembre" , "December")
+            .build();
+    private static Map<String, String> spanishMonths = ImmutableMap.<String,String>builder()
+            .put("enero",  "January")
+            .put("febrero", "February")
+            .put("marzo" ,"March")
+            .put("abril" , "April")
+            .put("mayo"	, "May")
+            .put("junio" , "June")
+            .put("julio"	, "July")
+            .put("agosto" , "August")
+            .put("septiembre" , "September")
+            .put("octubre" , "October")
+            .put("noviembre" , "November")
+            .put("diciembre" , "December")
+            .build();
+    private static Map<String, String> germanMonths = ImmutableMap.<String,String>builder()
+            .put("januar",  "January")
+            .put("februar", "February")
+            .put("märz" ,"March")
+            .put("april" , "April")
+            .put("mai"	, "May")
+            .put("juni" , "June")
+            .put("juli"	, "July")
+            .put("august" , "August")
+            .put("september" , "September")
+            .put("oktober" , "October")
+            .put("november" , "November")
+            .put("dezember" , "December")
+            .build();
 
     public static Date stringToDate(String input) {
         Date date = null;
@@ -73,5 +183,133 @@ public class DateUtil {
         } catch (ParseException e) {
         }
         return null;
+    }
+
+    private static String replaceMonths(Map<String, String> months, String dateStr) {
+
+        for(Map.Entry<String, String> entry : months.entrySet()) {
+            dateStr = dateStr.replace(entry.getKey(), entry.getValue());
+        }
+
+        return dateStr;
+    }
+
+    private static Optional<LocalDate> getJChronic (String date) {
+        Span span = Chronic.parse(date, opts);   // Year is treated as 1938 because of PointerType.PAST.
+        Optional<LocalDate> result = Optional.empty();
+        if (span != null) {
+            LocalDate localDate = Instant.ofEpochMilli(span.getBegin()*1000).atZone(ZoneId.of("GMT")).toLocalDate();
+            result = Optional.of(localDate);
+        }
+        return result;
+    }
+
+    private static Optional<LocalDate> getPretty (String date) {
+        List<Date> dates = prettyParser.parse(date);
+        Optional<LocalDate> result = Optional.empty();
+
+        if(!dates.isEmpty()) {
+            LocalDate localDate = dates.get(dates.size()-1).toInstant().atZone(ZoneId.of("GMT")).toLocalDate();
+
+            if(localDate.isAfter(LocalDate.now())) {
+                localDate = dates.get(0).toInstant().atZone(ZoneId.of("GMT")).toLocalDate();
+            }
+            result = Optional.of(localDate);
+        }
+        return result;
+    }
+
+    private static Optional<LocalDate> getNatty (String date) {
+
+        Optional<LocalDate> result = Optional.empty();
+
+        List<DateGroup> dateGroups = nattyParser.parse(date);
+        if(dateGroups.size() > 0) {
+            DateGroup dateGroup = dateGroups.get(dateGroups.size() - 1);
+            List<Date> ds = dateGroup.getDates();
+            Date d = ds.get(ds.size() - 1);
+            LocalDate localDate = d.toInstant().atZone(ZoneId.of("GMT")).toLocalDate();
+            result = Optional.of(localDate);
+
+        }
+
+        return result;
+    }
+
+    private static Optional<LocalDate> getDefault(String date) {
+        try {
+
+            Date d = stringToDate(date);
+            LocalDate localDate = d.toInstant().atZone(ZoneId.of("GMT")).toLocalDate();
+
+            return Optional.of(localDate);
+        } catch (IllegalArgumentException e) {
+
+            return Optional.empty();
+        }
+    }
+
+    public static String normaliseDate(String date) {
+
+        date = date.replaceAll("\\p{javaSpaceChar}+", " ").trim();
+
+        date = date.replaceAll("UTC\\+[0-9]+:[0-9]+","");
+
+        date = date.replaceAll("\\.","/");
+
+        date = date.replaceAll("-(?!\\p{Digit})", " ")
+                .replaceAll("(?<!\\p{Digit})-", " ");
+
+        date = date.replaceAll("(?<=\\p{Alpha}{2,})(?=\\p{Digit})", " ");
+        date = date.replace(",", " ");
+
+        date = replaceMonths(months,date);
+        date = replaceMonths(russianMonths,date.toLowerCase());
+        date = replaceMonths(spanishMonths,date.toLowerCase());
+        date = replaceMonths(germanMonths,date.toLowerCase());
+        date = replaceMonths(frenchMonths,date.toLowerCase());
+
+        return date;
+    }
+
+    private static Optional<LocalDate> resolveBestGuess(List<Optional<LocalDate>> guesses) {
+
+        List<LocalDate> dates = guesses.stream()
+                .filter( Optional::isPresent )
+                .map( Optional::get )
+                .filter( d -> d.isBefore(LocalDate.now()) )
+                .collect(Collectors.toList());
+
+        Optional<LocalDate> result = Optional.empty();
+
+        if(dates.size()==1) {
+            result = Optional.of(dates.get(0));
+        } else if (dates.size() > 1) {
+            Map<LocalDate, Long> counts = dates.stream().collect(
+                    groupingBy(Function.identity(), counting())
+            );
+
+            Optional<Map.Entry<LocalDate, Long>> favourite = counts.entrySet().stream().max(Map.Entry.comparingByValue());
+
+            if(favourite.isPresent()) {
+                result = Optional.of(favourite.get().getKey());
+            }
+        }
+
+        return result;
+    }
+
+    public static LocalDate getDate(String date) {
+
+        date = normaliseDate(date);
+
+        Optional<LocalDate> jchronicGuess = getJChronic(date);
+        Optional<LocalDate> prettyGuess = getPretty(date);
+        Optional<LocalDate> nattyGuess = getNatty(date);
+        Optional<LocalDate> defaultGuess = getDefault(date);
+
+        Optional<LocalDate> bestGuess = resolveBestGuess(ImmutableList.of(jchronicGuess, prettyGuess, nattyGuess, defaultGuess));
+
+        return bestGuess.orElseGet(()->null);
     }
 }
