+++ src/main/java/com/casm/acled/crawler/utils/Util.java
@@ -3,7 +3,7 @@ package com.casm.acled.crawler.utils;
 
 import com.casm.acled.camunda.BusinessKeys;
 import com.casm.acled.configuration.ObjectMapperConfiguration;
-import com.casm.acled.crawler.DateFilter;
+import com.casm.acled.crawler.IncorrectScraperJSONException;
 import com.casm.acled.dao.entities.ArticleDAO;
 import com.casm.acled.dao.entities.SourceDAO;
 import com.casm.acled.dao.entities.SourceListDAO;
@@ -11,16 +11,10 @@ import com.casm.acled.entities.article.Article;
 import com.casm.acled.entities.source.Source;
 import com.casm.acled.entities.sourcelist.SourceList;
 
-import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
-import com.joestelmach.natty.DateGroup;
-import com.joestelmach.natty.Parser;
-import com.mdimension.jchronic.Chronic;
-import com.mdimension.jchronic.Options;
-import com.mdimension.jchronic.tags.Pointer;
-import com.mdimension.jchronic.utils.Span;
+import com.google.common.io.Files;
 import org.camunda.bpm.spring.boot.starter.CamundaBpmAutoConfiguration;
 import org.camunda.bpm.spring.boot.starter.rest.CamundaBpmRestJerseyAutoConfiguration;
+import org.json.JSONObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -31,19 +25,17 @@ import org.springframework.boot.WebApplicationType;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Import;
-import org.ocpsoft.prettytime.nlp.PrettyTimeParser;
 
 
-import java.time.Instant;
+import java.io.File;
+import java.io.IOException;
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.nio.charset.Charset;
 import java.time.LocalDate;
-import java.time.ZoneId;
 import java.util.*;
-import java.util.function.Function;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
-import java.util.stream.StreamSupport;
-
-import static java.util.stream.Collectors.*;
 
 
 // We have to exclude these classes, because they only work in a web context.
@@ -54,6 +46,10 @@ import static java.util.stream.Collectors.*;
 @ComponentScan(basePackages={"com.casm.acled.dao"})
 public class Util implements CommandLineRunner {
     protected static final Logger logger = LoggerFactory.getLogger(Util.class);
+    // keyword query specific to potential articles of interest to ACLED
+//    public static String KEYWORDS2 = ".+(?:kill|killed|massacre|death|\\bdied\\b|\\bdead\\b|\\bbomb\\b|\\bbombed\\b|\\bbombing\\b|\\brebel\\b|\\battack\\b|\\battacked\\b|\\briot\\b|\\bbattle\\b|\\bprotest\\b|\\bclash\\b|\\bdemonstration\\b|\\bstrike\\b|\\bwound\\b|\\binjure\\b|\\bcasualty\\b|\\bdisplace\\b|\\bunrest\\b|\\bcasualties\\b|\\bvigilante\\b|\\btorture\\b|\\bmarch\\b|\\brape\\b).+";
+    public static String KEYWORDS =              ".*\\b(?:kill|massacre|death|died|dead|bomb|bombed|bombing|rebel|attack|attacked|riot|battle|protest|clash|demonstration|strike|wound|injure|casualty|displace|unrest|casualties|vigilante|torture|march|rape)\\b.*";
+//    private static final String matchingKeywords = "\\b(?:kill|massacre|death|died|dead|bomb|bombed|bombing|rebel|attack|attacked|riot|battle|protest|clash|demonstration|strike|wound|injure|casualty|displace|unrest|casualties|vigilante|torture|march|rape)\\b";
 
     @Autowired
     private ArticleDAO articleDAO;
@@ -64,233 +60,43 @@ public class Util implements CommandLineRunner {
     @Autowired
     private SourceDAO sourceDAO;
 
-    private static Options opts = new Options(Pointer.PointerType.PAST);
-    private static PrettyTimeParser prettyParser = new PrettyTimeParser();
-    private static PrettyTimeParser nonamericanPrettyParser = new PrettyTimeParser();
-
-    private static Parser nattyParser = new Parser();
-
-    private static Map<String, String> months = ImmutableMap.<String,String>builder()
-            .put("كَانُون ٱلثَّانِي",  "January")
-            .put("شُبَاط", "February")
-            .put("آذَار" ,"March")
-            .put("نَيْسَان" , "April")
-            .put("أَيَّار"	, "May")
-            .put("حَزِيرَان" , "June")
-            .put("تَمُّوز"	, "July")
-            .put("آب" , "August")
-            .put("أَيْلُول" , "September")
-            .put("تِشْرِين ٱلْأَوَّل" , "October")
-            .put("تِشْرِين ٱلثَّانِي" , "November")
-            .put("كَانُون ٱلْأَوَّل" , "December")
-
-            .put("يناير" , "January")
-            .put("فبراير" , "February")
-            .put("مارس" , "March")
-            .put("أبريل" , "April")
-            .put("إبريل" , "April")
-            .put("مايو" , "May")
-            .put("يونيو" , "June")
-            .put("يونية", "June")
-            .put("يوليو" , "July")
-            .put("يولية" , "July")
-            .put("أغسطس", "August")
-            .put("سبتمبر", "September")
-            .put("أكتوبر" , "October")
-            .put("نوفمبر" , "November")
-            .put("ديسمبر" , "December")
-            .build();
-
-    private static Map<String, String> russianMonths = ImmutableMap.<String,String>builder()
-            .put("января",  "January")
-            .put("февраля", "February")
-            .put("март" ,"March")
-            .put("апреля" , "April")
-            .put("май"	, "May")
-            .put("июня" , "June")
-            .put("июля"	, "July")
-            .put("августа" , "August")
-            .put("сентября" , "September")
-            .put("октября" , "October")
-            .put("ноября" , "November")
-            .put("декабря" , "December")
-            .build();
-
-    private static Map<String, String> frenchMonths = ImmutableMap.<String,String>builder()
-            .put("janvier",  "January")
-            .put("février", "February")
-            .put("mars" ,"March")
-            .put("avril" , "April")
-            .put("mai"	, "May")
-            .put("juin" , "June")
-            .put("juillet"	, "July")
-            .put("août" , "August")
-            .put("septembre" , "September")
-            .put("octobre" , "October")
-            .put("novembre" , "November")
-            .put("décembre" , "December")
-            .build();
-
-    private static Map<String, String> spanishMonths = ImmutableMap.<String,String>builder()
-            .put("enero",  "January")
-            .put("febrero", "February")
-            .put("marzo" ,"March")
-            .put("abril" , "April")
-            .put("mayo"	, "May")
-            .put("junio" , "June")
-            .put("julio"	, "July")
-            .put("agosto" , "August")
-            .put("septiembre" , "September")
-            .put("octubre" , "October")
-            .put("noviembre" , "November")
-            .put("diciembre" , "December")
-            .build();
-
-    private static Map<String, String> germanMonths = ImmutableMap.<String,String>builder()
-            .put("januar",  "January")
-            .put("februar", "February")
-            .put("märz" ,"March")
-            .put("april" , "April")
-            .put("mai"	, "May")
-            .put("juni" , "June")
-            .put("juli"	, "July")
-            .put("august" , "August")
-            .put("september" , "September")
-            .put("oktober" , "October")
-            .put("november" , "November")
-            .put("dezember" , "December")
-            .build();
-
-    private static String replaceMonths(Map<String, String> months, String dateStr) {
-
-        for(Map.Entry<String, String> entry : months.entrySet()) {
-            dateStr = dateStr.replace(entry.getKey(), entry.getValue());
-        }
-
-        return dateStr;
-    }
-
-    private static Optional<LocalDate> getJChronic (String date) {
-        Span span = Chronic.parse(date, opts);   // Year is treated as 1938 because of PointerType.PAST.
-        Optional<LocalDate> result = Optional.empty();
-        if (span != null) {
-            LocalDate localDate = Instant.ofEpochMilli(span.getBegin()*1000).atZone(ZoneId.of("GMT")).toLocalDate();
-            result = Optional.of(localDate);
-        }
-        return result;
-    }
-
-    private static Optional<LocalDate> getPretty (String date) {
-        List<Date> dates = prettyParser.parse(date);
-        Optional<LocalDate> result = Optional.empty();
-
-        if(!dates.isEmpty()) {
-            LocalDate localDate = dates.get(dates.size()-1).toInstant().atZone(ZoneId.of("GMT")).toLocalDate();
-
-            if(localDate.isAfter(LocalDate.now())) {
-                localDate = dates.get(0).toInstant().atZone(ZoneId.of("GMT")).toLocalDate();
-            }
-            result = Optional.of(localDate);
-        }
-        return result;
-    }
-
-    private static Optional<LocalDate> getNatty (String date) {
-
-        Optional<LocalDate> result = Optional.empty();
-
-        List<DateGroup> dateGroups = nattyParser.parse(date);
-        if(dateGroups.size() > 0) {
-            DateGroup dateGroup = dateGroups.get(dateGroups.size() - 1);
-            List<Date> ds = dateGroup.getDates();
-            Date d = ds.get(ds.size() - 1);
-            LocalDate localDate = d.toInstant().atZone(ZoneId.of("GMT")).toLocalDate();
-            result = Optional.of(localDate);
-
-        }
-
-        return result;
-    }
-
-    private static Optional<LocalDate> getDefault(String date) {
+    // Returns the originating domain of a given url - minus any trailing 'www'
+    public static String getDomain(String urlString)  {
         try {
-
-            Date d = DateUtil.stringToDate(date);
-            LocalDate localDate = d.toInstant().atZone(ZoneId.of("GMT")).toLocalDate();
-
-            return Optional.of(localDate);
-        } catch (IllegalArgumentException e) {
-
-            return Optional.empty();
+            URL url = new URL(urlString);
+            return (url.getHost().startsWith("www")) ? url.getHost().substring(4) : url.getHost();
+        } catch (MalformedURLException e){
+            throw new RuntimeException(e);
         }
     }
 
-    public static String normaliseDate(String date) {
-
-        date = date.replaceAll("\\p{javaSpaceChar}+", " ").trim();
-
-        date = date.replaceAll("UTC\\+[0-9]+:[0-9]+","");
-
-        date = date.replaceAll("\\.","/");
-
-        date = date.replaceAll("-(?!\\p{Digit})", " ")
-                .replaceAll("(?<!\\p{Digit})-", " ");
-
-        date = date.replaceAll("(?<=\\p{Alpha}{2,})(?=\\p{Digit})", " ");
-        date = date.replace(",", " ");
-
-        date = replaceMonths(months,date);
-        date = replaceMonths(russianMonths,date.toLowerCase());
-        date = replaceMonths(spanishMonths,date.toLowerCase());
-        date = replaceMonths(germanMonths,date.toLowerCase());
-        date = replaceMonths(frenchMonths,date.toLowerCase());
-
-        return date;
-    }
-
-
-    private static Optional<LocalDate> resolveBestGuess(List<Optional<LocalDate>> guesses) {
-
-        List<LocalDate> dates = guesses.stream()
-                .filter( Optional::isPresent )
-                .map( Optional::get )
-                .filter( d -> d.isBefore(LocalDate.now()) )
-                .collect(Collectors.toList());
-
-        Optional<LocalDate> result = Optional.empty();
-
-        if(dates.size()==1) {
-            result = Optional.of(dates.get(0));
-        } else if (dates.size() > 1) {
-            Map<LocalDate, Long> counts = dates.stream().collect(
-                    groupingBy(Function.identity(), counting())
-            );
-
-            Optional<Map.Entry<LocalDate, Long>> favourite = counts.entrySet().stream().max(Map.Entry.comparingByValue());
-
-            if(favourite.isPresent()) {
-                result = Optional.of(favourite.get().getKey());
-            }
+    // Processes a M52 job json to scraper rules
+    public static String processJobJSON(String json) throws IncorrectScraperJSONException {
+        JSONObject jobj = new JSONObject(json);
+        try {
+            return jobj.getJSONArray("components").getJSONObject(0).getJSONObject("opts").getJSONArray("fields").toString();
+        } catch (Exception e) {
+            throw new IncorrectScraperJSONException();
         }
-
-        return result;
     }
 
-    public static LocalDate getDate(String date) {
-
-        date = normaliseDate(date);
-
-        Optional<LocalDate> jchronicGuess = getJChronic(date);
-        Optional<LocalDate> prettyGuess = getPretty(date);
-        Optional<LocalDate> nattyGuess = getNatty(date);
-        Optional<LocalDate> defaultGuess = getDefault(date);
-
-        Optional<LocalDate> bestGuess = resolveBestGuess(ImmutableList.of(jchronicGuess, prettyGuess, nattyGuess, defaultGuess));
-
-        return bestGuess.orElseGet(()->null);
+    public static String processScraperJSON(String json){
+        // BUG FOUND - NEED TO USE job.json!!
+        return null;
     }
 
-//    public LocalDate simplDateFallback(String dateString) {
+    // returns a web scraper based on a job spect of last_scrape file
+    public static String processJSON(File scraperLocation) throws IOException, IncorrectScraperJSONException {
+        String json = Files.asCharSource(scraperLocation, Charset.defaultCharset()).read();
+        return (scraperLocation.getName().equals("last_scrape.json")) ? processScraperJSON(json) : processJobJSON(json);
+    }
+
+//    public static void main(String[] args) throws Exception {
+//        getDomain("http://www.0.com:5000");
+//    }
+
+
+    //    public LocalDate simplDateFallback(String dateString) {
 //
 //    }
 
@@ -312,7 +118,7 @@ public class Util implements CommandLineRunner {
             LocalDate localDate = null;
             try {
 
-                localDate = getDate(text[1]);
+                localDate = DateUtil.getDate(text[1]);
             } catch (RuntimeException e) {
                 continue;
             }
@@ -342,14 +148,13 @@ public class Util implements CommandLineRunner {
         }
     }
 
-    private static final String matchingKeywords = "\\b(?:kill|massacre|death|died|dead|bomb|bombed|bombing|rebel|attack|attacked|riot|battle|protest|clash|demonstration|strike|wound|injure|casualty|displace|unrest|casualties|vigilante|torture|march|rape)\\b";
 
     private void deleteNonMatchingArticles() {
-        Pattern pattern = Pattern.compile(matchingKeywords);
+        Pattern pattern = Pattern.compile(KEYWORDS);
         for (Article article : articleDAO.getAll()) {
             String text = article.get(Article.TEXT);
 
-            if(!pattern.matcher(text).find()) {
+            if(!pattern.matcher(text).matches()) {
                 System.out.println("REMOVE " + text);
                 articleDAO.delete(article);
             } else {
