+++ src/test/java/uk/ac/susx/tag/norconex/TestDateFilter.java
@@ -1,7 +1,6 @@
 package uk.ac.susx.tag.norconex;
 
-import com.casm.acled.crawler.DateFilter;
-import com.casm.acled.crawler.utils.Util;
+import com.casm.acled.crawler.utils.DateUtil;
 import org.junit.Test;
 
 public class TestDateFilter {
@@ -11,7 +10,7 @@ public class TestDateFilter {
 
         String date = "fl  10/06/2019 wjmm";
 //        DateFilter df = new DateFilter();
-        System.out.println(Util.getDate(date));
+        System.out.println(DateUtil.getDate(date));
 
     }
 
@@ -20,7 +19,7 @@ public class TestDateFilter {
 
         String date = "09/02/2019";
 //        DateFilter df = new DateFilter();
-        System.out.println(Util.getDate(date));
+        System.out.println(DateUtil.getDate(date));
 
     }
 }
