+++ src/test/java/uk/ac/susx/tag/norconex/TestJSONParser.java
@@ -1,9 +1,8 @@
 package uk.ac.susx.tag.norconex;
 
 import com.casm.acled.crawler.IncorrectScraperJSONException;
-import com.casm.acled.crawler.utils.Utils;
+import com.casm.acled.crawler.utils.Util;
 import org.apache.commons.io.FileUtils;
-import org.json.JSONObject;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -21,7 +20,7 @@ public class TestJSONParser {
         String f = "/Users/jp242/Documents/Projects/ACLED/ManualScrapers/rinascrapers/azatliqorg/job.json";
         try {
             String json = FileUtils.readFileToString(new File(f), Charset.defaultCharset());
-            String processed = Utils.processJSON(new File(f));
+            String processed = Util.processJSON(new File(f));
             System.out.println(processed);
 //            String parsed = "[{\"field\":\"root/root\",\"tags\":[{\"custom\":\"#content\"}]},{\"field\":\"field.name/title\",\"tags\":[{\"tag\":\"h1\",\"class\":\"pg-title\"}]},{\"field\":\"field.name/date\",\"tags\":[{\"tag\":\"div\",\"class\":\"published\"}]},{\"field\":\"field.name/article\",\"tags\":[{\"tag\":\"div\",\"class\":\"wsw\"}]}]\n";
 //            Assert.assertEquals(processed,parsed);
