+++ src/test/java/uk/ac/susx/tag/norconex/TestWordBoundaries.java
@@ -1,6 +1,6 @@
 package uk.ac.susx.tag.norconex;
 
-import com.casm.acled.crawler.utils.Utils;
+import com.casm.acled.crawler.utils.Util;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -14,13 +14,13 @@ public class TestWordBoundaries {
 
         String shouldMatch = "A bomb has exploded outside of quatar. All were killed";
         String shoudNotMatch = "a nonbomb has pexploded outside of quatar. All were notkilled.";
-        Assert.assertTrue(Pattern.matches(Utils.KEYWORDS,shouldMatch));
-        Assert.assertFalse(Pattern.matches(Utils.KEYWORDS,shoudNotMatch));
+        Assert.assertTrue(Pattern.matches(Util.KEYWORDS,shouldMatch));
+        Assert.assertFalse(Pattern.matches(Util.KEYWORDS,shoudNotMatch));
 
         String shouldDemoMatch = "Protestors performed a demonstration outside parliament this morning.";
         String shouldNotDemoMatch = "Protestorsires redemonstration outside parliament this morning.";
-        Assert.assertTrue(Pattern.matches(Utils.KEYWORDS,shouldDemoMatch));
-        Assert.assertFalse(Pattern.matches(Utils.KEYWORDS,shouldNotDemoMatch));
+        Assert.assertTrue(Pattern.matches(Util.KEYWORDS,shouldDemoMatch));
+        Assert.assertFalse(Pattern.matches(Util.KEYWORDS,shouldNotDemoMatch));
 
     }
